refactor: kebab-case file naming + enforce version bump on every PR
File renames (camelCase/PascalCase -> kebab-case):
- cli: loadPlugins.ts
- shared/manifest-parsers: hls.ts, hls.test.ts, parser.ts (+ snapshots) reorganized under hls/
- shared/plugins/caption-extractor/cea: cea608*, cea708*, dtvcc*, ceaDecoder
- shared/utils: arrayUtils, buildRawText, createEmptyRegion, createView,
  hexToUint8, textConstants, uint8ToString, mp4/dataViewReader, mp4/parser,
  cea/getParsedCaption, cea/getStreamAndLanguages
- viewer: AppHeader, ChangelogPopover, PrimaryButton, FilesystemWriter,
  app/components/ui/copy-button -> components/ui/, app/lib -> lib/
Imports updated across cli/, shared/, viewer/ to match.

CI policy:
- pr.yaml: changelog job now passes --require-bump, so every PR must
  increment root package.json version.
- push.yaml: removed the check-version gate and the --release changelog
  step; release job runs unconditionally on every push to main and reads
  the version directly from package.json for the tag and viewer build.
- scripts/check-changelog.mjs: new --require-bump flag that fails when
  the current version equals the base ref's version.

Docs:
- README.md: Contributing section reflects "every PR bumps + every merge
  releases" policy.
- CHANGELOG.md: housekeeping entry for the policy change.
