+++ cli/src/plugins/load-plugins.ts
