+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cmdt",
-  "version": "0.5.1",
+  "version": "0.5.2",
   "description": "Common Media Diagnostic Tool",
   "repository": {
     "type": "git",
