+++ shared/src/drm/pssh.ts
@@ -1,5 +1,5 @@
 import { type ILogObj, Logger } from "tslog";
-import Mp4Parser from "../utils/mp4/parser.js";
+import Mp4Parser from "../utils/mp4/mp4-parser.js";
 import type { ParsedBox, Pssh } from "../utils/mp4/types.js";
 import { type PlayreadyData, PlayreadyParser } from "./playready/playready.js";
 import { type WidevineData, WidevineParser } from "./widevine/widevine.js";
