+++ shared/src/manifest-parsers/hls/hls-parser.ts
