+++ shared/src/plugins/gap-checker/gap-checker.ts
@@ -1,6 +1,6 @@
 import { type Manifest, MediaType, type Representation, type Segment } from "../../manifest.js";
 import type { Report } from "../../report.js";
-import Mp4Parser from "../../utils/mp4/parser.js";
+import Mp4Parser from "../../utils/mp4/mp4-parser.js";
 import type { Mdhd, ParsedBox, Tfdt, Tfhd, Tkhd } from "../../utils/mp4/types.js";
 import { secondsToMilliseconds } from "../../utils/time-utils.js";
 import { Plugin, type PluginArtifact } from "../plugin.js";
