+++ shared/src/utils/build-raw-text.ts
