+++ shared/src/utils/create-empty-region.ts
