+++ shared/src/utils/hex-to-uint8.ts
