+++ shared/src/utils/text-constants.ts
